Add two ansible variables to configure how env vars are deployed
This configure the ansible playbook provides two ways for a user
to deploy the ".env" which contains the configuration for signalboost.

**deploy_method**

The default settings is "copy" which will copy a local file to the server.
This file by default is .env but can be changed using the variable **deploy_file**

If deploy_method is set to "blackbox" two things happen:

* Ansible deploys GPG keys
* Ansible deploys .env using blackbox on the remote server